ci-operator: Update the dockerfile_path for the OLM and operator-registry images
Update the upstream OLM and operator-registry ci-operator/config prow
configurations and revert a change made by the auto-config-brancher
ensuring that ART and CI dockerfile(s) names match ocp-build-data.

This enforcement only applies to promoted images, and we're now
promoting ocp images from the downstream monorepo.

This also should be a no-op as that `dockerfile_path: Dockerfile`
already exists in both of the repositories in the root directory, and we
symbolically linked the content in those files, to the
`operator-*.Dockerfile` equivalent around a week ago as a temporary
workaround.