Move scope of coreos namespace to all clusters
To enable building of CoreOS images on any cluster via Gangplank we need
to move the cluster scope from boot02 to all the clusters.

The ci-operator has also been granted admin scope permissions to the
coreos namespace. This will allow us to use the ci-operator to create
pods in the cores namespaces via Prow jobs.

Signed-off-by: Ben Howard <ben.howard@redhat.com>