ci-operator: Remove the prow configuration for the upstream operator-registry component
Remove the prow configuration for the release-4.8+ branches in the
operator-framework/operator-registry component.

That unit test prow job is a duplicate as there's an existing GH action
in the upstream repository that covers that same check.

Remove that integration/e2e check with the latest OCP OLM images as we
can transistion that work to be a GH action at a later date.