ci-operators/jobs: Mark the OCP e2e-related jobs as required
Update the existing downstream OLM e2e-related prow jobs as required
checks. Previously these were marked as optional in order to push
through content as we built up the repository in the 4.8 release. Now
that we've transistion to building and promoting from this repository,
mark these e2e tests as required checks.