periodics: adjust interval on machine-os-content promote jobs
The interval settings on our promote jobs were out of whack when
compared to what we need.  Specifically, releases that were older and
not updated as frequently were bing promoted every 15m, but the
release under active development (4.8) was only geting promoted every
4h.

This changes the interval on released versions (4.5, 4.6, 4.7) to be
24h and increases the frequency on the actively developed version
(4.8) to be 15m.