hypershift: Add a HyperShift cluster launch workflow
This commit adds a new `hypershift-launch-aws` workflow which can run tests in
the context of an ephemeral HyperShift cluster hosted on AWS. See the workflow
documentation for more details.

Also introduced is a new openshift/release periodic which runs the OCP
conformance tests using the new HyperShift workflow. This job is the basis for a
forthcoming cluster-bot integration.

Everything introduced here is for experimental use only and shouldn't be relied
upon for anything critical. For now, there are no guarantees any of it will work
at all. We need to commit these basic pieces to continue iteration towards a
supportable state.