CoreOS: add CoreOS BuildConfig and permissions
CoreOS Assembler has the ability to run as a custom build strategy.
Since CoreOS images are neither containers or buildable as such standard
plumbing through Prow does not work. By running the builds as a custom
build strategy, the CoreOS team will be able to use Prow for basic PR
tests and gating logic and perform basic tests in a public fashion.

Signed-off-by: Ben Howard <ben.howard@redhat.com>