Report nightly master failure to OCS Slack channel
OCS engineers should be notified when the nightly master job
fails. At present the only explicitly supported option for this
in ci-operator is Slack. While suboptimal, this is at least a
good stopgap measure. GitHub reporting is also an option in
Prow; as Slack reporting is a simple passthrough to Prow at
the ci-operator level, this may be a viable replacement in the
near future.

Signed-off-by: egafford <egafford@redhat.com>