ci-operator/step-registry/openshift/e2e/test: Get GCP project from the cluster
With e8788a7d04 (create a cluster profile for the
openshift-gce-devel-ci-2 gcp account, 2021-04-15, #17770), we're
starting to shard between two GCP accounts, so instead of hard-coding
the project, pull it out of the cluster's infrastructure resource [1].

I dunno if we can rely on having jq around.  If we can, we could make
a single 'oc' call to get the infrastructure object, and then extract
the various properties locally.  But just calling 'oc' again for each
property isn't a big deal, so that's what I've gone with here.

The gcloud setup might appear to be unused, but is actually part of
the local context that the test suite might require.  Background in
b4e60e4994 ([gcp] Add gcloud sdk to test container for gcp e2e tests,
2019-08-06, #4643).

[1]: https://github.com/openshift/api/blob/38058be7d6ef61b980f52f3633c41bbf8e20a0ea/config/v1/types_infrastructure.go#L392