cvp: add a fast probe job & step
Add a job to be used as a fast probe that OpenShift CI works. The job
can be triggered by `cvp-trigger`, consumes identical paramaters, but
does not install a testing cluster. It touches (via `oc image info`) the
provided images using the CVP image proxy credentials and creates an
artifact at well-known location in artifact directory. The artifact can
be used to verify that the `cvp-trigger` params were correctly
propagated to the job.