Support proxy in gather step commands
This change looks for a proxy-conf.sh file in the shared directory, and
if present sources it.  Baremetal IPI runs in an isolated bubble on a
remote host, so `oc` from CI won't work there without punching a hole
through to the virtualized environment. We will do that with a squid
proxy.

This change is independent from the baremetal IPI changes, as touching
gather triggers too many jobs, and after several attempts none of them
triggered any metal job to test if this works.