[cincinnati] Enable linting for master tests
the master tests currently do not have
rustfmt and yaml-lint tests configured.
this commit adds these tests to the master.
In order to achieve this, we need to create a custom
container with rustup installed. This is being done by
creating an src-base image which has all the tools that we
need for linting. this container image will only be used
for linting purpose.

Given that rust-toolset ubi image will be retired in some
time, changing src-base to use ubi and installing rustup, another
motivation for this change is that the ubi image has a rust
binary installed which cannot be used to change the rust version.
We are installing rustup to achieve this.

additional files created by running
make jobs and
make ci-operator-config