splunk-forwarder-operator: *Almost* standardize prow config
This brings splunk-forwarder-operator's prow jobs in line with the
[boilerplate standards](https://github.com/openshift/boilerplate/blob/master/boilerplate/openshift/golang-osd-operator/README.md#prow),
except that the `vuln-check-*` jobs are preserved, as these are not yet
covered by boilerplate (see OSD-7146). In order to preserve these jobs,
we make the following deviations from the standard config:
- We keep the ubi-minimal `base` image because it is the basis for
building the splunk-forwarder and splunk-heavyforwarder images.
- Instead of building the operator image to `unused`, we build it to
`splunk-forwarder-operator` so it can be used by `vuln-check-operator`.

OSD-7087