ci-operator/config/openshift/release: Add 4.7.z CI to 4.7.1 rollback tests
Sometimes folks get into difficulties during updates, and within
updates, rolling back to a previous patch release should work.  To
keep an eye on this, and also to discourage backporting features into
4.7.z that might break rollbacks, run a periodic job to test 4.7.1 ->
candidate-nightly -> 4.7.1 rollbacks.  The random aborts should get us
all sorts of timings, which is good because users might get jumpy and
want to go back to A at any point during their A->B update.  The
downside of random aborts is that we will not get as much coverage on
late-in-update components like machine-config.

We went GA with 4.7.0, but I'm treating 4.7.1 as the first GA, because
4.7.1 got [1], and without that 4.7.later -> 4.7.0 rollbacks hang like
this 4.7.8 -> 4.7.0 rollback [2]:

  Cluster did not complete upgrade: timed out waiting for the condition: Cluster operator ingress is degraded
  Operator degraded (IngressControllersDegraded): Some ingresscontrollers are degraded: ingresscontroller "default" is degraded: DegradedConditions: One or more other status conditions indicate a degraded state: CanaryChecksSucceeding=False (CanaryChecksRepetitiveFailures: Canary route checks for the default ingress controller are failing)

The release format I've used to define oldest-supported-z-rollback is
documented in [3].

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1932649#c6
[2]: https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/release-openshift-origin-installer-launch-gcp/1385456065000247296
[3]: https://docs.ci.openshift.org/docs/architecture/ci-operator/#testing-with-an-existing-openshift-release