openshift-e2e: support hypershift cluster types
Update the OCP e2e test step to recognize the hypershift cluster type
and treat it the same as AWS for now, which is the only place hypershift
clusters are currently offered.

This hack is necessary because cluster type is currently inseparable
from cluster profile (https://issues.redhat.com/browse/DPTP-1988).