gcp: Fix shortened instance group for newer installer versions
The bootstrap instance groups were using the suffix 'bootstrap-instance-group'
in their names in previous versions whici was taking up the 63 character
limit for the names. This was fixed in the installer.

Adding the fix in the gcp tests to use the older name first for older version
support and if it fails, use the newer, smaller suffix.