remove 4.6 vsphere ovn presubmit job
This is the logical conclusion to this PR [0] which made the
perma-failing job not be gating requiring an /override. The
job fails because of docker rate limiting and since it never
passes, it's useless. I'm told that backporting the fixes
to 4.6 that would get past the image pull rate limiting are
not easy and not worth the trouble to figure out. So let's
trash this job completely.

[0] https://github.com/openshift/release/pull/17743

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>