gather: support overriding image for disconnected environments
Prior to 4.8, must-gather tries to access the internet on disconnected
environments. This changes adds support for the steps that run
must-gather to support reading a must-gather-image.sh file from the
shared directory to override which image to use.

Baremetal disconnected will populate this file with the locally hosted
image when working on a release older than 4.8.