Test assisted service using merged source
Previously this step used the assisted-test-infra image as a base which
made it difficult to get access to the correct source for the PR under
test. Specifically this would end up with an image built using a PR
merge commit into master, but the deploy would happen using the PR head.

This would cause a problem if the PR was not rebased on top of another
change that significantly altered the deploy process.

This PR switches the base image for the step to get access to the source
of the merge commit then pulls the test infra source straight from the
github master branch.