Openstack: Remove unused KURYR_ENABLED from step registry.
The env variable KURYR_ENABLED is not used anywhere in the step registry.
But listing it under the env variables of a step makes it a requirement that
it be defined somewhere by the user.
Even though it has a default value if the user does not define it explicitely,
leaving it around could cause problems for users in the future who would be confused
as to why they have to give it a value (define it).