cri-o: Enable kata-containers-2 CI
Let's enable, for now not always running, a basic kata-containers CI
running against the `main` branch of the project.

Once we ensure this is stable enough on the kata-containers side, we can
switch it to run on every PR automatically.

The job used on the kata-containers side is:
http://jenkins.katacontainers.io/job/kata-containers-2-crio-PR/

It's important to note that we'll only run those tests on the following
branches on the CRI-O side:
- `master`
- `release-1.21` onwards

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>