hypershift: fix inputs to cluster destroy command
Before this commit, the infra ID for a cluster was incorrectly specified on the
call to `destroy cluster`, which in some unusual circumstances could cause infra
cleanup to fail if the hostedcluster itself disappeared before the infra
teardown attempt.

This commit fixes the input so that infra will be reliably destroyed even if the
hostedcluster has vanished before calling the destroy command.