Parameterize fields in operator scorecard script
The optional-operators-cvp-common-scorecard-commands.sh script is
used by CVP for testing of operator bundles in external test
environments.  However, two of the hard-coded fields in the script,
an auth flag and that path to the scorecard config, need to be
configurable in order to work in other environments.  The default
values have been set to the currently hard-coded ones.

Signed-off-by: Ryan Cole <rycole@redhat.com>