ci-operator/jobs/openshift/machine-config-operator: Don't run e2e, etc. on docs/
The machine-config operator started using this approach in in
4e51092c2a (ci-operator/jobs/openshift/machine-config-operator: skip
docs/ changes, 2019-01-14, #2576).  In this commit, I'm extending it
to cover all the formerly-always-run e2e and other slow jobs, because
folks shouldn't have to say "nah, not worth waiting out CI" when they
are contemplating a docs improvement that has no impact on shipped
code.

I haven't bothered to extend to unit tests and other fast stuff,
because they're cheap.