Remove RELEASE_IMAGE_LATEST from assisted presubmits
We set the latest release for 4.8 before we had it as part of the
defaults on assisted-service repository on
https://github.com/openshift/assisted-service/blob/master/data/default_ocp_versions.json

Now that we maintain a stable version, we would want all assisted
presubmit jobs to run with a stable version (since nightlies might not
be stable).
Periodics will continue to run with the latest version.