Update MCO e2e-metal-ipi job definitions
In [0], the configuration for metal-ipi jobs was changed to be more
consistent. While we've been able to run for a while without those
changes, recently there were some more changes to how we filter out
tests that aren't expected to pass on certain metal jobs. I suspect
those changes were not compatible with our old job definitions, so
we now have tests running that can never pass.

This updates e2e-metal-ipi to be ipv4, consistent with the other
metal-associated repos. It also adds the e2e-metal-ipi-ovn-ipv6
job if we want to test ipv6 functionality for patches.

0: https://github.com/openshift/release/commit/0f313f6c04a16bb62fe4e44cd1f9d1e796d3cd01#diff-5160099cf0dfc7df93b1bab410e543c423d60aa5b7552b9f5ef31bfedf555d68