Configure real-time workers with tuned profile
The real-time kernel designed to run an application with the minimal possible latency and
not designed to run a lot of short-living processes, result in a lot of system or user processes
can have CPU starvation, and the environment will have unexpected failures during containers start and stop procedures.

To improve the situation under our real-time CI lane, the CI will deploy the real-time tuned profile that
should configure workers with a set of parameters that should prevent CPU starvation of system or user processes.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>