step-registry: Fix unbound variable error on optional-operators-ci-subscribe-commands.sh
On optional-operators/ci/subscribe/optional-operators-ci-subscribe-commands.sh
the CVS variable is not set if an install plan is not found, however that variable is later
referenced which makes the script to fail with "CSV: unbound variable" error.

This fixes that problem by initializing the CSV variable at the begin of the script.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>