Add a distinct branch protector job for the openshift org
This change adds a second branchprotector job that only protects the
openshift org and uses the merge robot token. The main branch protector
job in contrast will not protect the openshift org anymore.

This is because we only have limited tokens in the GH app for the
openshift org but have plenty of them on the merge robot account.