cvp: replace rehearsal defaults with proper multi-stage params
This change is a step towards obsoleting the custom parameter passing
interface between cvp-trigger and ci-operator. We start with replacing
the "rehearsal defaults" by multi-stage parameters where values are
defined in tests (this has a side benefit of allowing each job to have a
separate set of good defaults).

The actual values are (for now!) only used when `OO_*` values are not
set (`OO_*` values are passed by the custom parameter code). Next step
is to make `cvp-trigger` override the values of the multi-stage params
with the same values it passes to `OO_*` params. Then we can change the
steps to actually use the multi-stage params instead of the custom ones.
Then we can remove the custom code from both `cvp-trigger` and
`ci-operator`.