Add job to refresh all BZs on master/main
When we open master/main branches for the development of the next OCP
version, we need to recalculate validity of all in-flight PRs BZ links.
Many of them already have `valid-bug` that was put there because the PR
is filed for a bug in the *previous* version. After code freeze, only
bugs targeting the next version should be considered valid.