Add openshift-e2e-openstack-csi workflow to csi-driver-manila-operator
This patch will allow to run the first functional job
(openshift-e2e-openstack-csi workflow) on the
csi-driver-manila-operator repo, using the new OSP cloud (named
openstack-vh-mecha)

It also adds the necessary bits to load OC_CACERT if a certificate is
present. Our new cloud has self signed certificate so this is required.