Hive: Get rid of unnecessary hivebase
The Dockerfile.hivebase wasn't doing anything useful. (It was installing
openssh-clients -- but the Dockerfile does that anyway.) So collapse the
image builds to get rid of it. This should let us get rid of the
Dockerfile.hivebase in the hive repo, and make the appsre ci.ext job
redundant as well.