hypershift: increase e2e test timeouts to improve stability
Contention within the CI account seems to lead to a wild variance times to
create or even sometimes destroy cloud infrastructure.

This commit increases the hypershift e2e test timeout to be long enough to
smooth out even the worst of these issues. This is so that we have a more stable
test pipeline at the expense of getting early warnings about potential
performance regressions. We're hoping to start collecting and measuring metrics
for these sorts of tests in a way that doesn't make us reliant on this
particular timeout to do that sort of analysis (as the SLIs/SLOs should be
codified external to this test definition.)