Hive: run_if_changed => skip_if_only_changed
Previously we were skipping jobs for PRs that only changed files
matching `^docs/`. We were doing this via `run_if_changed`, which is
inclusive, so we had to use an awkward inverted regex.

Using the new `skip_if_only_changed`, which is *exclusive*, we are now
able to supply a more powerful regex to skip PRs which only change
paths:
- starting with `docs/`
- ending with `.md`
- whose basename is any of
  - `.gitignore`
  - `OWNERS`
  - `PROJECT`
  - `LICENSE`