Hive: wrap e2e in generic-claim, use 4.7
Since cutting over to `cluster_claim` (#19405), the e2e job was
- having problems starting pods, probably (see below) due to missing
[pull permissions](https://steps.ci.openshift.org/reference/ipi-install-rbac)
[Later] Actually it was due to [BZ
1972383](https://bugzilla.redhat.com/show_bug.cgi?id=1972383). We'll
temporarily use 4.7 to avoid this. (It's what we were using before
anyway.)

and

- not `gather`ing anything, so when it failed, we couldn't tell what was
going wrong. It looks like the [generic-claim
workflow](https://steps.ci.openshift.org/workflow/generic-claim) is
designed for exactly this, so this commit adds it.

HIVE-1567