Hive: Use `cluster_claim` for AWS e2e (attempt 2)
PR #19405 and #19443 tried to cut hive's AWS e2e test over to using
`cluster_claim`. We're doing that again, but addressing several of the
issues we encountered in and since those PRs:

1. Nothing was being gathered for post-mortem use. This commit includes
`workflow: generic-claim` to do that.
2. We were trying to use 4.8, but it was broken for a couple of reasons
([BZ 1971540](https://bugzilla.redhat.com/show_bug.cgi?id=1971540)
(upstream [1972383](https://bugzilla.redhat.com/show_bug.cgi?id=1972383))
and DPTP-2262). This commit uses 4.7 (which is what we were using before
anyway).
3. With ClusterPools, we no longer get access to the CI environment's
AWS credentials. We've created an account on our team cluster and put
the credentials into Vault. This commit injects those creds into the
test environment and uses them to create the creds file in the location
expected by the test. In the future we'll refactor the test to accept
them from the environment instead, so this commit also sets the
appropriate environment variables.
4. Likewise the CI cluster's pull secret is no longer provided
implicitly, so inject it explicitly.
5. Likewise the CI cluster's SSH keypair is no longer provided. Generate
one.
6. The default domain used by the test was based on the CI environment's
AWS account. Now that we're using ours, override it accordingly.

HIVE-1567