Onboard ExternalDNS Operator
This commit adds preliminary configs generated via
`make new-repo` for adding the new
[ExternalDNS Operator](https://github.com/openshift/external-dns-operator)
repo to OpenShift CI.

This commit configures a basic unit test job for the ExternalDNS
operator as a starting point to get provide basic CI and merge
control for the ExternalDNS operator repo.

This is in support of https://issues.redhat.com/browse/NE-453.