openstack: Remove confusing "minimal" workflow
The current "openshift-e2e-openstack-minimal" workflow does not run a
minimal set of tests: it does not run ANY test. It just installs.

To avoid further confusion (and unnecessary duplication), this change
removes this "minimal" workflow.

Jobs that actually need not to run any test are modified to use the
regular OpenStack workflow "openshift-e2e-openstack-ipi" and to override
the test step with a no-op step.