ci-operator/step-registry/gather/aws-console: UTF-8 for output
In recent CI, we have failed console gathers like [1]:

  Gathering console logs for i-026ba0637214cd027
  'ascii' codec can't encode character '\u2026' in position 14093: ordinal not in range(128)
  Failed to gather console logs

This is descended from aa31ea376d (gather/aws-console: Make log
gathering non-fatal, 2021-04-14, #17735).  But we want to gather these
logs, not silently fail to gather the logs.  This commit restores our
old "fail the step when we fail to gather" behavior, and sets
PYTHONIOENCODING [2] to ask for UTF-8 instead of ASCII output.

[1]: https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/logs/periodic-ci-openshift-[…]e2e-aws-serial/gather-aws-console/build-log.txt
[2]: https://docs.python.org/3/using/cmdline.html#envvar-PYTHONIOENCODING