WIP: ci-operator/jobs/openshift/release: Set owner/description labels for 4.9-e2e-aws-canary
Testing this approach to see if its feasible.  Using comma-delimited
GitHub usernames for owners.  Probably want to think through
the label keys a bit more before we actually do this.

I think these are actually a better fit for annotations, because we
don't need to use them in list queries.  But release CI doesn't seem
to allow annotations [1].

[1]: https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_release/19929/pull-ci-openshift-release-master-generated-config/1411044173171134464#1:build-log.txt%3A20