Implement basic proxy CI for OpenStack platform
This effort is a first iteration over testing different scenarios, all
related to "Bring Your Own Network" feature.

This patch creates 2 new kind of steps (each of them have provision
& deprovision steps):

* One for creating a network in OpenStack, that will be used to deploy
  the OCP cluster. It's called `machinesubnet`. It creates a network,
  a subnet, a router and plug the subnet onto that router so it can
  reach the external networks (Internet).
* One for creating the bastion node, which in fact will act as a proxy.
  This will use Squid and basically help the CI to reach the previously
  created internal network, to deploy and test OCP.
  Note that the Baremetal team is also working on this topic and we will
  probably revisit our steps / scripts to align how we do.

Then, this patch creats a new workflow, called
`openshift-e2e-openstack-proxy` and sets the CONFIG_TYPE to `proxy`.
We will have more config types in the future, for disconnected installs
for example.

And finally, create a job that will run on both the installer and CAPO
which will always run the proxy job and be optional (for now) to merge
patches.

Note: the jobs don't run the test suite because a lot of test don't
support HTTP proxy and therefore fail. Until we find a sane solution,
for now we're satisfied by just deploying an OCP cluster and if it
finishes successfully then the job will pass.

Implements OSASINFRA-2022