ci-operator/config: Update the o/operator-framework-olm unit test command(s)
Update the o/operator-framework-olm ci-operator configs for master/4.9+
branches and ensure that the XDG_CACHE_HOME environment variable is set
to a tmp directory due to the controller-runtime changes introduced in https://github.com/kubernetes-sigs/controller-runtime/commit/58c17f696263fc4a35ef5de198f99aa1717e7b87.

Before, this was defaulting to ${HOME}/.cache, which was producing
permission denied errors in the unit prow job when attempting to create
that cache directory. This is because the unit tests don't have root
permissions.