Fix failure to timeout in ccm-conf-apply-feature-gate-commands.sh
The primary issue here is the use of the --foreground argument to
timeout in waitFor(), which states:

  when not running timeout directly from a shell prompt, allow COMMAND
  to read from the TTY and get TTY signals; in this mode, children of
  COMMAND will not be timed out

Combined with an unnecessary subshell in execute() this meant that the
executed command would never be signalled to terminate by timeout.

There's more to it, though, because the primary shell did still
correctly exit with a non-zero status. Perhaps the job continued to
run due to job control in the supervisor.