reorder install step traps
The traps were added to make "Preemptions" jobs deprovision step successful - make sure that metadata.json file is moved even if the container was destroied in the middle.
We recently saw that some Preemptions jobs don't cleanup after them self and timeout due to grace period ending.
I suspect that due to the wait in the job killing we are never reaching the "prepare_next_steps" function.
This change replaces the order of traps making sure we first prepare the next steps and only after we destroy child processes.