openstack upgrades: cluster_profile resource names
The OS_CLOUD value that was in the config of the openstack jobs,
`openstack-cloud`, is only valid for legacy job-wide secrets. OpenStack
step-registry tests now use `openstack` as the value in all instances,
because each cloud is saved on its own `clouds.yaml` that is dynamically
selected in the openstack-conf-cloud step, based on the cluster_profile.

The correct external network name and the correct flavor name for the
cloud are injected autmatically by openstack-conf-resourcenames. They
are also not necessary in the config any more.