Add ability to fail if core dumps are found
only turning this on for 4.9+ ovn upgrade jobs for now. Those
jobs in 4.8 and older do see core dumps from ovsdb-server and
ovn northd. That was fixed [0] in 4.9+ and this will help
prevent those specific cases from creeping back in. This can
be expanded to other jobs, platforms (only aws now) and versions
after some time with just aws-ovn-upgrade jobs running with it.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1943363

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>