Bug 1985960 - oVirt 4.8 tests are failing on resize
This change adds steps to generate the csi test manifest from a step instead of from the secrets.
It adds a step for release 4.6-4.8 which doesn't include the resizing tests and a step for 4.9 and above which do support resizing.
We also remove the prepare env step since it is not needed anymore