Add Proxy CI for OpenStack platform
This commit includes a new Job for a proxy
installation with a restricted Network. The
bastion VM is configured with two networks,
one with access to the external world (Bastion network) and
another without (Machines Network). The installer will use the
restricted Machines network and rely on the Proxy
for external access.

The job will run on demand from the installer PRs and also every 72h in
periodic.

Co-Authored-By: Maysa Macedo <maysa.macedo95@gmail.com>
Co-Authored-By: Emilien Macchi <emilien@redhat.com>