Make sure TPNU or CustomNoUpgrade FG does not affect test results
Either TPNU or CustomNoUpgrade results in alerts firing in e2e tests,
however it is expected in CCM use-case and should not make the test fail.

- `Managed cluster should start all core operators` - kube-api-server operator
has a specific alert, notifying the cluster is non-upgradable with TP or CNU
featureGates. This outputs:

```
Upgradeable=False FeatureGates_RestrictedFeatureGates_CustomNoUpgrade: FeatureGatesUpgradeable: "CustomNoUpgrade" does not allow updates
```

marking operator non-healthy in e2e tests. This is not critical and
is expected in successfull job pass.