Add more doc for TEST_SKIPS
When some tests fail, sometimes Prow don't display the exact name of the
test and a space might be missing.
Now we document that it's advised to run 'openshift-tests run --dry-run (...)'
to make sure that the tests we want to skip have the right name and we
build a REGEX that will work.