Hive: Increase claim timeouts
We started seeing some e2e[-pool] jobs fail because the clusterclaim
wasn't getting a CD before the configured timeout. This is probably
because the pool doesn't have a high enough maxSize/maxConcurrent (see
DPTP-2450). Even assuming they fix that, it behooves us to allow a bit
longer for the claim to be fulfilled.

This commit sets the claim timeout for e2e and e2e-pool based on the
overall job timeout of 4h, minus the anecdotal observed time the jobs
seem to take (~1.5h for e2e, ~1h for e2e-pool), minus a buffer of ~30m
for comfort.