ci-operator: Replace --timeout with --request-timeout for rsh calls
The --timeout approach landed back in ad2dbe0b5b (Dump iptables on
every node during teardown, 2019-01-24, #2652).  But oc dropped the
deprecated --timeout option in openshift/oc@8931142584 (Remove
deprecated --timeout flag from oc rsh, 2021-01-25, openshift/oc#631).
That change landed in 4.7 and was backported to 4.6.23 [1], but not to
4.5.

I've confirmed with Dan Winship that we still want the iptables saves,
and don't just want to drop these blocks.

Generated with:

  $ sed -i 's/rsh --timeout=20/--request-timeout=20s rsh/' $(git --no-pager grep -l 'rsh.*timeout' ci-operator)

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1926172#c7