ci-operator/step-registry/gather/extra: Set 10 minute timeout
Looking at successful gathers over the past day, this will give us
plenty of headroom:

  $ curl -s 'https://search.ci.openshift.org/search?maxAge=24h&type=build-log&context=0&search=Step+.*-gather-extra+succeeded+after' | jq -r 'to_entries[].value | to_entries[].value[].context[]' | sed 's/.*succeeded after \(\([0-9]*\)m\)\?\(\([0-9]*\)s\)[.]/\2 * 60 + \4/;s/^ [*] 60 [+] //' | while read LINE; do echo "$((${LINE}))"; done | sort -n | uniq -c
        1 20
        6 30
       12 40
        5 50
       22 60
        1 61
       56 70
       66 80
       70 90
      331 100
      407 110
        1 111
      261 120
      121 130
       66 140
       24 150
       14 160
       10 170
        3 180
        6 190
        1 200
        1 210
        1 220
        2 230
        1 250
        1 1180

So 10m is over twice the largest successful gather, with the exception
of that one 1180s job.  Loking at that job:

  $ curl -s 'https://search.ci.openshift.org/search?maxAge=24h&type=build-log&context=0&search=Step+.*-gather-extra+succeeded+after+19m40s' | jq -r 'keys[]'
  https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/kubevirt_hyperconverged-cluster-operator/1484/pull-ci-kubevirt-hyperconverged-cluster-operator-main-okd-hco-e2e-upgrade-index-gcp/1427685715176591360

It's not entirely clear to me why the gather step took so long there,
but I expect we would certainly be able to gather enough useful
information in 10m to help with debugging, and I don't think waiting
that step out in that case is worth letting the gather run longer if
there's some hiccup in the script that causes it to block (because we
want to be able to get through and run other cleanup steps before the
job times us out).