Add e2e-metal-assisted-multi-version job
On https://github.com/openshift/assisted-service/pull/2289 the 4.8
version became the default.
There's no need to set OPENSHIFT_VERSION=4.8 in many places anymore as
it is the default.

assisted-setup step was changed to be aligned with the default on
assisted-service repository according to data/default_ocp_versions.json
file.

A new job was added named e2e-metal-assisted-multi-version that will be
used to run all the supported versions. It will be triggered as a
periodic job and as well when the "data" directory changes.

From now on, presubmit jobs will run only the stable default version.

ocm-2.3 branch will remain the same since the default wasn't changed
there.