prow-monitoring: create alerts for release-controller bugzilla errors
This PR adds a prometheus alert that fires when a bugzilla error is
recorded by the release-controller and sends the alert to the
`#forum-release-controller` channel. Currently, the time is configured
for `1m`, which means any bugzilla error will trigger an alert. If the
alert gets triggered to often by flakes, it can be changed later.