Declare assisted-setup POST_INSTALL_COMMANDS env var
The new `POST_INSTALL_COMMANDS` defines a list of commands
to run once the installation has finished.

If the env var is non-empty, set it to
`${SHARED_DIR}/assisted-post-install.sh` and copy it to the remote
Packet machine in under the same name.
After the installation has finished, in case the file exist, source it
and execute its content.