alerts: change release-controller bugzilla error alert
The `rate` function in prometheus returns the per second increase rate
of a vector. This would result in an error only causing the query to
change from `0 to `0.04`. To fix this, we can either change the query to
be `> 0` or use `increase` instead of `rate`. This PR does the former.