Update o/operator-framework-olm build_root configuration for master/4.10+
Update the build_root configuration and read from the .ci-operator.yaml
from the repository instead of specifying the build root information in
the release repository. This namely allows us to make changes to the
build root Dockerfile, and have those changes tested in the same PR.
Previously, changes made to the build root Dockerfile will only be run
once they land in master as it's reading from the HEAD version.