repos: Point advanced-virt to the "8.4 EUS" CDN
The repository currently used has no distinction between the AV version
of the packages that are present there.

This has never been an issue as OpenShift was moving to latest RHEL
release as soon as the release was out, which changed for 4.8, 4.9, and
4.10, which will stick to RHEL-8.4 as its base os for RHCOS.

However, with OpenShift sticking to a specific RHCOS release we need to
ensure new QEMU packages are not mistakenly consumed.  The proper way to
do so is using the EUS CDN for RHEL-AV 8.4 as, for sure, packages for a
different (as in newer) version of RHEL won't be present there.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>