Ensure Proxy works on installations with Kuryr
During a proxy installation two Networks are configured
for the bastion VM, one gives access to the internet and
another to the OCP cluster. Since Pods running on Pod Network
runs on a separate Neutron Network, a route needs to be
configured at bastion VM specifying the correct interface
that should handle that traffic.